refactor(desktop): extract subscription ipc
